requirements.txt
requests
urllib3<2.0.0
python-dotenv
python-dateutil
streamlit
